cdp-agent/requirements.txt
# Unpinned requirements: Let pip resolve all versions automatically for maximum compatibility with cdp-sdk, coinbase-agentkit, and other dependencies.
# If you need to pin versions for production, do so after confirming compatibility.
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asn1crypto
async-timeout
attrs
bcl
bip-utils
bitarray
cbor2
cdp-sdk
certifi
cffi
charset-normalizer
ckzg
click
coincurve
coinbase-agentkit
crcmod
cryptography
cytoolz
distro
ecdsa
ed25519-blake2b
egcd
eth-account
eth-hash
eth-keyfile
eth-keys
eth-rlp
eth-typing
eth-utils
eth_abi
fastapi
frozenlist
h11
hexbytes
httpcore
httpx
httpx-sse
idna
jiter
jsonpatch
jsonpointer
langchain
langchain-core
langchain-openai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-sdk
langsmith
msgpack
multidict
nilql
openai
orjson
packaging
pailliers
parsimonious
propcache
py-sr25519-bindings
pycparser
pycryptodome
pydantic
pydantic_core
PyJWT
PyNaCl
python-dateutil
python-dotenv
pyunormalize
PyYAML
rabinmiller
regex
requests
requests-toolbelt
rlp
six
sniffio
SQLAlchemy
starlette
tenacity
tiktoken
toolz
tqdm
types-requests
typing_extensions
urllib3
uvicorn
web3
websockets
yarl
zstandard
flask
# Use coinbase-agentkit and compatible cdp-sdk. Removed cdp-agentkit-core and cdp-langchain due to conflicts.
# Removed pip and setuptools; manage these with your Python environment, not as dependencies